requirements.txt
djangorestframework-simplejwt==5.1.0
django==3.2.0
djangorestframework
psycopg2-binary==2.9.3
drf-yasg
django-filter
# pytest==7.1.1
# pytest-cov
# pytest-django==3.7.0

